Created ncurses example, ncurses_getline()
Created small test example of a program making use of the (not-so-standard-C-but-still-very-common) ncurses library, for extering greater control over command line interfaces, to learn how to incorporate it into the chat program. The chat program will use the library to seperate the input and output and prevent the ugliness of the former UI (where, because of the way output was handled, it could interrupt the user typing at any moment). Because input is tied so closely to the screen, a new getline() (which was used extensively in client.c) was needed for use with ncurses. Because ncurses has no built-in getline(), a function was written to handle this, *ncurses_getline(), and it appears to work.