Refined ncurses_getline()
Refined the ncurses_getline() function to behave more similarly to the POSIX getline(). It now takes in a char** variable and uses this to 'return' the value of the line. The allocated size of the array by the end of the function is also returned using a pointer to a size_t variable. Finally, the return value was used to indicate the number of characters actually read (including the newline). One remaining thing that should be done is to check for null upon reallocations.