Partially fixed communication between the client and server
Fixed teh communication between the client and server in regards to initially choosing username. Is now impossible for client to overrun the servers buffer or to leave extra characters in the servers read queue after it finishes reading what it believed to the be the username. Started the process of preventing array overflow in creating messages with buffer-sized usernames and any-sized messages. Fixed the sending of "[User] joined the chatroom!" to only send to the other users. Finally, for real, fixed only allowing one user with a username. (The problem was that, in the while loop entered after picking a username, an extra newline was being sent.)