Fixed client/server communication, other issues
Fixed the communication between the client and the server so that now the user on the client side can send messages of any length without causing overflows and memory errors on the server side. Fixed an issue with not assigning the value of *output to the tmp realloc variable in ncurses_getline(). Updated the makefile for runNcurses so, for the time being, only the node on the tux cluster being used must be specified.