Move command processing to client side.
To reduce the amount of computation the server has to do, and because the server implementation of the commands was 'broken' by buffer overflows anyway, just moved the processing of commands to the client side, which also required reformatting regular messages. Basically, now, the client formats all of the strings before sending and the server just collects the strings and redistributes them. Right now, the roll and me commands work, and whisper should be good soon.